+++ src/grim.nim
@@ -2,10 +2,14 @@ import tables
 
 export tables
 
+import grim/entities
+import grim/paths
 import grim/box
 import grim/io
 import grim/graph
 
+export entities
+export paths
 export box
 export io
 export graph