+++ Changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.3.0 [2020-12-30]
+
+- Filter when iterating nodes and edges
+- PathCollections to find similar paths under a common API (`navigate`, `step`, `steps`, and `follow` procs)
+- Paths to easily walk the graph
+- Basic Neo4j client, use `dump` proc to load graph database via http as grim LPG
+- Works with nim 1.4.2 (not orc yet due to dependence on NimYAML)
+
 ## v.0.2.0 [2020-02-07]
 
 - Complete "Northwind" tutorial to show how SQL database is transferred to LPG.
@@ -16,7 +24,7 @@
 - Make sure entites are ref objects to avoid excessive copying.
 - Updated all documentation.
 - Equality operators for nodes and edges.
-- Documentation website at https://ebran.github.io/grim/.
+- Documentation website at [Github](https://ebran.github.io/grim/).
 
 ### v.0.1.1 [2020-01-12]